Refactored config into its own module
Fixes OMN-11
