+++ src/lib.rs
@@ -23,62 +23,16 @@ extern crate log;
 extern crate web3;
 
 use blockchain::{Blockchain, Kind};
-use std::env;
+use config::Config;
 use std::error::Error;
 
 mod blockchain;
-
-// Environment variables and their defaults
-const ENV_ORIGIN_ADDRESS: &str = "MOSAIC_ORIGIN_ADDRESS";
-const ENV_AUXILIARY_ADDRESS: &str = "MOSAIC_AUXILIARY_ADDRESS";
-const DEFAULT_ORIGIN_ADDRESS: &str = "http://127.0.0.1:8545";
-const DEFAULT_AUXILIARY_ADDRESS: &str = "http://127.0.0.1:8546";
-
-/// Global config for running a mosaic node.
-pub struct Config {
-    /// Address of the origin chain, e.g. "127.0.0.1:8485"
-    origin_address: String,
-    /// Address of the auxiliary chain, e.g. "127.0.0.1:8486"
-    _auxiliary_address: String,
-}
-
-impl Config {
-    /// Reads the configuration from environment variables and creates a new Config from them. In
-    /// case an environment variable is not set, a default fallback will be used.
-    pub fn new() -> Config {
-        // Read origin address from env and set it or fallback to default
-        let origin_address = env::var(ENV_ORIGIN_ADDRESS);
-        let origin_address = match origin_address {
-            Ok(address) => address,
-            Err(_) => {
-                info!("No origin chain address given, falling back to default.");
-                DEFAULT_ORIGIN_ADDRESS.to_string()
-            }
-        };
-        info!("Using origin chain address: {}", origin_address);
-
-        // Read auxiliary address from env and set it or fallback to default
-        let auxiliary_address = env::var(ENV_AUXILIARY_ADDRESS);
-        let auxiliary_address = match auxiliary_address {
-            Ok(address) => address,
-            Err(_) => {
-                info!("No auxiliary chain address given, falling back to default.");
-                DEFAULT_AUXILIARY_ADDRESS.to_string()
-            }
-        };
-        info!("Using auxiliary chain address: {}", auxiliary_address);
-
-        Config {
-            origin_address,
-            _auxiliary_address: auxiliary_address,
-        }
-    }
-}
+pub mod config;
 
 /// Runs a mosaic node with the given configuration.
 /// Prints all accounts of the origin blockchain to std out.
 pub fn run(config: Config) -> Result<(), Box<Error>> {
-    let blockchain = Blockchain::new(Kind::Eth, config.origin_address);
+    let blockchain = Blockchain::new(Kind::Eth, config.origin_address().to_string());
 
     let accounts = blockchain.get_accounts();
 
@@ -88,35 +42,4 @@ pub fn run(config: Config) -> Result<(), Box<Error>> {
     }
 
     Ok(())
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-
-    #[test]
-    fn the_config_reads_the_environment_variables() {
-        let config = Config::new();
-        assert_eq!(config.origin_address, DEFAULT_ORIGIN_ADDRESS.to_string());
-        assert_eq!(
-            config._auxiliary_address,
-            DEFAULT_AUXILIARY_ADDRESS.to_string()
-        );
-
-        env::set_var(ENV_ORIGIN_ADDRESS, "10.0.0.1");
-        let config = Config::new();
-        assert_eq!(config.origin_address, "10.0.0.1");
-        assert_eq!(
-            config._auxiliary_address,
-            DEFAULT_AUXILIARY_ADDRESS.to_string()
-        );
-
-        env::set_var(ENV_AUXILIARY_ADDRESS, "10.0.0.2");
-        let config = Config::new();
-        assert_eq!(config.origin_address, "10.0.0.1");
-        assert_eq!(config._auxiliary_address, "10.0.0.2");
-
-        env::remove_var(ENV_ORIGIN_ADDRESS);
-        env::remove_var(ENV_AUXILIARY_ADDRESS);
-    }
 }
